requirements.txt
PyQt6==6.3.1
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
google-api-python-client==2.86.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
openai==0.28.0
keyring==24.2.0
cryptography==41.0.3
ultralytics==8.0.196
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
tqdm>=4.65.0 